+++ CHANGELOG.md
@@ -8,10 +8,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Added
+- Risk-review: Added settings: idStart (starting point for line drawing), textPosY (vertical text position: center\top\bottom), textPosX (horizontal text position: center\default), textOffset (horizontal text offset from line), textColor
 
 ### Changed
 
 ### Fixed
+- Risk-review: Fixed incorrect processing of empty values for captions
 
 ## [1.134.1]