+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Risk-review: Added settings: idStart (starting point for line drawing), textPosY (vertical text position: center\top\bottom), textPosX (horizontal text position: center\default), textOffset (horizontal text offset from line), textColor
 
 ### Changed
+- Updated logic for changing the title of the browser tab
 
 ### Fixed
 - Risk-review: Fixed incorrect processing of empty values for captions